Introduce base class Handler for both events and commands handling
