+++ eventsourcing_helpers/handler.py
@@ -1,0 +1,12 @@
+from typing import Callable
+
+from eventsourcing_helpers.serializers import from_message_to_dto
+
+
+class Handler:
+    handlers: dict = {}
+
+    def __init__(self, message_deserializer: Callable=
+                 from_message_to_dto) -> None:  # yapf: disable
+        assert self.handlers
+        self.message_deserializer = message_deserializer
